requirements.txt
backcall==0.2.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.42.0
django-cors-headers==3.3.0
django-extensions==2.2.9
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
oauthlib==3.1.0
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
Pygments==2.6.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.4
